% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fit_cp_model}
\alias{fit_cp_model}
\title{Fit Change Point}
\usage{
fit_cp_model(count_data, low, high, model = "lm", plot = TRUE, level = 0.95)
}
\arguments{
\item{count_data}{A dataset of visit counts, must contain variables 'days_since_dx' and
'n'. Where 'days_since_dx' is the number of days since diagnosis (negative values
before diagnosis) and 'n' are the number of visits}

\item{low}{the low value in the training window (i.e., the point closest to
diagnosis to include in the training window)}

\item{high}{the high value in the training window (i.e., the point furthest from
diagnosis to include in the training window)}

\item{model}{the model to be used. Current options include "lm", "exp", "lm_period",
"exp_period"}
}
\description{
Fit Change Point
}
\examples{

fit_cp_model(count_data = visit_counts,
             low = 40,high = 180,
             model = "lm",
             plot = FALSE)

}
