% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fit_cp_profile}
\alias{fit_cp_profile}
\title{Generate Change Point Profile}
\usage{
fit_cp_profile(
  count_data,
  low,
  high,
  models = c("lm", "exp", "lm_period", "exp_period", "auto_arima"),
  levels = c(0, 0.9, 0.95),
  window_grid = NA
)
}
\arguments{
\item{count_data}{A dataset of visit counts, must contain variables 'days_since_dx' and
'n'. Where 'days_since_dx' is the number of days since diagnosis (negative values
before diagnosis) and 'n' are the number of visits}

\item{low}{the low value in the training window (i.e., the point closest to
diagnosis to include in the training window)}

\item{high}{the high value in the training window (i.e., the point furthest from
diagnosis to include in the training window)}

\item{models}{the set of models}

\item{levels}{the set of levels}
}
\description{
Generate Change Point Profile
}
