% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ssd_curve}
\alias{plot_ssd_curve}
\title{Plot SSD curve}
\usage{
plot_ssd_curve(count_data, max_window = NA)
}
\arguments{
\item{count_data}{A dataset of visit counts, must contain variables 'days_since_dx' and
'n'. Where 'days_since_dx' is the number of days since diagnosis (negative values
before diagnosis) and 'n' are the number of visits}
}
\description{
Plot SSD curve
}
\examples{

plot_ssd_curve(count_data = visit_counts)

}
